train/requirements.txt
python==3.9.18
tensorflow==2.15.0
sklearn==1.5.0
numpy==1.26.4
pandas==2.2.2
torch==2.3.1
torchvision==0.18.1
tensorcircuit==0.12.0
torchsummary==1.5.1
flask==3.0.3
keras-swa
